refactor: tools folder
