+++ src/main/kotlin/org/jetbrains/research/testspark/tools/TestProcessor.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testspark.tools.generatedTests
+package org.jetbrains.research.testspark.tools
 
 import com.gitlab.mvysny.konsumexml.konsumeXml
 import com.intellij.openapi.components.service
@@ -16,9 +16,6 @@ import org.jetbrains.research.testspark.services.LLMSettingsService
 import org.jetbrains.research.testspark.services.PluginSettingsService
 import org.jetbrains.research.testspark.services.TestsExecutionResultService
 import org.jetbrains.research.testspark.settings.llm.LLMSettingsState
-import org.jetbrains.research.testspark.tools.LibraryPathsProvider
-import org.jetbrains.research.testspark.tools.TestCompilerFactory
-import org.jetbrains.research.testspark.tools.ToolUtils
 import java.io.File
 import kotlin.io.path.Path
 import kotlin.io.path.createDirectories
